+++ ForgeMenuChoice/ForgeMenuChoice/docs/Changelog.md
@@ -4,6 +4,7 @@
 #### Todo
 [] Document the tooltip override feature + how to replace the textures in Content Patcher.
 [] Make sure all strings I want translated are complete, and ask for translations.
+[] Handle splitscreen. PerScreen things where needed.
 
 #### Known issues
 * Controllers probably won't be able to lock onto the arrows properly, owing to the....nontraditional menu setup here. It's likely not something I can fix, either.
@@ -14,6 +15,7 @@
 * Added tooltips: hover over the name of the enchantment to get a description. (You may need to find a certain journal scrap first.)
 * Slightly adjust the menu for longer enchantment names in some languages. Fix issue where the enchantment names would not change if you switched languages.
 * Made this mod's textures targetable in Content Patcher.
+* Splitscreen fixes.
 
 #### Version 1.0.0
 * Initial upload.